Services: Moved InactiveDateCalculator
